chore: replace #[allow(unused)] by `#[cfg(unix)]` when appropriate
Biggest change is splitting platform-specific code from permissions.rs
(similarly to flags.rs in the same directory).
