+++ src/fs/fields.rs
@@ -21,18 +21,19 @@
 #![allow(clippy::struct_excessive_bools)]
 
 /// The type of a file’s group ID.
+#[cfg(unix)]
 pub type gid_t = u32;
 
 /// The type of a file’s inode.
-#[allow(unused)]
+#[cfg(unix)]
 pub type ino_t = u64;
 
 /// The type of a file’s number of links.
-#[allow(unused)]
+#[cfg(unix)]
 pub type nlink_t = u64;
 
 /// The type of a file’s user ID.
-#[allow(unused)]
+#[cfg(unix)]
 pub type uid_t = u32;
 
 /// The type of user file flags
@@ -68,6 +69,7 @@ impl Type {
 /// The file’s Unix permission bitfield, with one entry per bit.
 #[derive(Copy, Clone)]
 #[rustfmt::skip]
+#[cfg(unix)]
 pub struct Permissions {
     pub user_read:      bool,
     pub user_write:     bool,
@@ -116,6 +118,7 @@ pub struct PermissionsPlus {
 
 /// The permissions encoded as octal values
 #[derive(Copy, Clone)]
+#[cfg(unix)]
 pub struct OctalPermissions {
     pub permissions: Permissions,
 }
@@ -126,7 +129,7 @@ pub struct OctalPermissions {
 /// multiple directories. However, it’s rare (but occasionally useful!) for a
 /// regular file to have a link count greater than 1, so we highlight the
 /// block count specifically for this case.
-#[allow(unused)]
+#[cfg(unix)]
 #[derive(Copy, Clone)]
 pub struct Links {
     /// The actual link count.
@@ -139,7 +142,7 @@ pub struct Links {
 /// A file’s inode. Every directory entry on a Unix filesystem has an inode,
 /// including directories and links, so this is applicable to everything exa
 /// can deal with.
-#[allow(unused)]
+#[cfg(unix)]
 #[derive(Copy, Clone)]
 pub struct Inode(pub ino_t);
 
@@ -156,12 +159,12 @@ pub enum Blocksize {
 
 /// The ID of the user that owns a file. This will only ever be a number;
 /// looking up the username is done in the `display` module.
-#[allow(unused)]
+#[cfg(unix)]
 #[derive(Copy, Clone)]
 pub struct User(pub uid_t);
 
 /// The ID of the group that a file belongs to.
-#[allow(unused)]
+#[cfg(unix)]
 #[derive(Copy, Clone)]
 pub struct Group(pub gid_t);
